refactor(Separator): 문자열을 분리하는 계산기의 로직을 분리
계산기가 문자열을 분리하지 않고 구분자가 문자열을 분리하도록 한다.
이에 Separator 클래스를 생성한다.
