+++ src/main/java/calculator/domain/Calculator.java
@@ -1,38 +1,13 @@
 package calculator.domain;
 
 public class Calculator {
-    private static final String defaultSeparator = "[,:]";
-    private String customSeparator = "";
-
+    private Separator separator;
     private Operands operands;
 
     public Calculator(String inputValue) {
-        setCustomSeparator(inputValue);
-
-        String calcValue = getCalcValue(inputValue);
-        String[] splitInput = splitBySeparator(calcValue);
-
+        Separator separator = new Separator(inputValue);
+        String[] splitInput = separator.splitBySeparator(inputValue);
         this.operands = new Operands(splitInput);
-    }
-
-    private void setCustomSeparator(String inputValue) {
-        if (inputValue.startsWith("//")) {
-            this.customSeparator = String.valueOf(inputValue.charAt(2));
-        }
-    }
-
-    private String getCalcValue(String inputValue) {
-        if (inputValue.startsWith("//")) {
-            return inputValue.substring(5);
-        }
-        return inputValue;
-    }
-
-    private String[] splitBySeparator(String inputValue) {
-        if (customSeparator.isEmpty()) {
-            return inputValue.split(defaultSeparator);
-        }
-        return inputValue.split(customSeparator);
     }
 
     public int sum() {
